Switch Sprockets extension to be a new-style Middleman::Extension
