+++ lib/middleman-sprockets.rb
@@ -2,5 +2,5 @@ require "middleman-core"
 
 Middleman::Extensions.register(:sprockets) do
   require "middleman-sprockets/extension"
-  Middleman::Sprockets
+  Middleman::SprocketsExtension
 end
